minor fixes

+++ rl4echo/datamodules/RL_3d_datamodule.py
@@ -127,7 +127,7 @@ class RL3dDataset(Dataset):
 
         return {'img': img.type(torch.float32),
                 'gt': mask.type(torch.LongTensor) if self.allow_real_gt or self.test else torch.zeros_like(torch.tensor(mask)),
-                'approx_gt': approx_gt.type(torch.float32),
+                'approx_gt': approx_gt.type(torch.LongTensor),
                 'use_gt': torch.tensor(self.use_gt[idx]),
                 'image_meta_dict': {'case_identifier': self.df.iloc[idx]['dicom_uuid'],
                                     'original_shape': original_shape,
